Remove streaming file and move relevant code to pos_final_state.
